requirements.txt
absl-py==1.0.0
anyio==3.7.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
atomicwrites==1.4.0
attrs==23.1.0
audioread==3.0.0
autopep8==1.6.0
backcall==0.2.0
backports.weakref==1.0.post1
beautifulsoup4==4.10.0
bkcharts==0.2
black==19.10b0
bleach==6.0.0
boto==2.49.0
boto3==1.26.103
botocore==1.29.103
Bottleneck==1.3.2
brotlipy==0.7.0
bvh-converter==1.0.2
bvhtoolbox==0.1.3
cachetools==5.0.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
comtypes==1.1.9
contextlib2==0.6.0.post1
cycler==0.10.0
cytoolz==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dpcpp-cpp-rt==2022.0.3
entrypoints==0.4
et-xmlfile==1.0.1
exceptiongroup==1.1.2
fastcache==1.1.0
filelock==3.6.0
fonttools==4.31.2
future==0.18.2
gdown==4.7.1
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h5py==2.10.0
HeapDict==1.0.1
helpdev==0.7.1
huggingface-hub==0.4.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
importlib-resources==5.12.0
iniconfig==1.1.1
intel-cmplr-lib-rt==2022.0.3
intel-cmplr-lic-rt==2022.0.3
intel-opencl-rt==2022.0.3
intel-openmp==2022.0.3
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
jdcal==1.4.1
jedi==0.18.2
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.1.0
json5==0.9.5
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
kiwisolver==1.4.0
librosa==0.9.2
llvmlite==0.39.1
locket==0.2.1
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==3.0.1
mkl==2022.0.3
mkl-fft==1.3.0
mpmath==1.2.1
multipledispatch==0.6.0
mypy-extensions==0.4.3
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.6.3
nlpaug @ git+https://github.com/makcedward/nlpaug.git@23800cbb9632c7fc8c4a88d46f9c4ecf68a96299
nltk==3.7
notebook==6.5.4
notebook_shim==0.2.3
numba==0.56.2
numexpr==2.8.1
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
open3d==0.9.0.0
opencv-python==4.7.0.72
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.7.0
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.7.0
pkgutil_resolve_name==1.3.10
pluggy==0.13.1
ply==3.11
pooch==1.6.0
praat-parselmouth==0.4.1
progressbar2==4.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.39
protobuf==3.19.4
psutil==5.9.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycosat==0.6.3
pycparser==2.21
pydub==0.25.1
Pygments==2.11.2
pyparsing==3.0.7
pyreadline==2.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.3
python-dateutil==2.8.2
python-utils==3.1.0
pytorch-pretrained-bert==0.6.2
pytz==2022.1
PyWavelets==1.3.0
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
qtconsole==5.2.2
QtPy==1.9.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.4.0
rsa==4.8
s3transfer==0.6.0
sacremoses==0.0.49
scikit-image==0.18.1
scikit-learn==1.0.2
scikit-learn-extra==0.2.0
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
simplegeneric==0.8.1
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
statsmodels==0.12.2
tables==3.6.1
tbb==2021.5.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.1
terminado==0.17.1
testpath==0.6.0
TextGrid==1.5
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.5
tokenizers==0.9.2
toml==0.10.2
toolz==0.11.2
torch==1.11.0+cu113
torch-tb-profiler==0.4.0
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==3.4.0
transforms3d==0.4.1
typed-ast==1.5.2
typing_extensions==4.1.1
unicodecsv==0.14.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zipp==3.7.0
zope.event==4.5.0